requirements.txt
Django==1.9.1
funcsigs==0.4
gunicorn==19.4.5
mock==1.3.0
pbr==1.8.1
Pillow==3.1.0
six==1.10.0
wheel==0.26.0
